cluster: fix race condition setting suicide prop
There is no guarantee that the `suicide` property of a worker in the master
process is going to be set when the `disconnect` and `exit` events are emitted.
To fix it, wait for the ACK of the suicide message from the master before
disconnecting the worker.
Modify `test-regress-GH-3238` so it checks both the `kill` and `disconnect`
cases. Also take into account that the `disconnect` event may be received after
the `exit` event.