cluster: call remove worker functions only once
In the current implementation, when `worker.disconnect` is called from
the master, the `removeHandlesForWorker` and `removeWorker` functions
are called multiple times: once in the `disconnect` method and then
once or twice in the `exit` and `disconnect` event handlers of the
worker process. This was causing the test `test-cluster-disconnect` to
fail with this error:

assert.js:89
  throw new assert.AssertionError({
  ^
AssertionError: Resource leak detected.
    at removeWorker (cluster.js:321:7)
    at ChildProcess.<anonymous> (cluster.js:378:28)
    at ChildProcess.g (events.js:264:16)
    at emitNone (events.js:68:13)
    at ChildProcess.emit (events.js:167:7)
    at finish (internal/child_process.js:657:14)
    at nextTickCallbackWith0Args (node.js:448:9)
    at process._tickCallback (node.js:377:13)

The problem being the handler created in `worker1` had been removed in
the call to the `disconnect` method; later on, a handler with the same
key had been created with `worker2`; then the `disconnect` event from
`worker1` fired before calling `worker2.disconnect` causing the
assertion.

To fix this add a `masterDisconnected` property to make sure the
`remove` functions aren't called in the event listeners if they have
already been called in `worker.disconnect`.