child_process: fix handleless NODE_HANDLE handling
It can happen that `recvmsg()` may return an error on ancillary data
reception when receiving a `NODE_HANDLE` message (for example
`MSG_CTRUNC`). This would end up, if the handle type was `net.Socket`,
on a `message` event with a non null but invalid `sendHandle`. Avoid
this by checking the `handle` validity before performing the
`handleConversion`.